v1.2.0: refactor codebase
