+++ src/framework/render.js
@@ -1,5 +1,7 @@
-import App from '../components/App';
+let Component, Target;
 
-export default function renderApp() {
-  document.getElementById('app-root').innerHTML = ` ${App()} `;
+export default function renderApp(componentFunction, targetElementId) {
+  if (componentFunction) Component = componentFunction;
+  if (targetElementId) Target = targetElementId;
+  document.getElementById(Target).innerHTML = ` ${Component()} `;
 }
